workshop-online/12-factor/api/requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.22
